requirements.txt
amqp
aniso8601
asgiref
async-timeout
billiard
blinker
CacheControl
cachetools
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
cron-descriptor
cryptography
dataclasses
Deprecated
diff-match-patch
dj-database-url
Django
django-advanced-filters
django-browser-reload
django-celery-beat
django-cron
django-environ
django-extensions
django-htmx
django-import-export
django-jazzmin
django-jinja
django-js-asset
django-meta
django-phonenumber-field
django-phonenumbers
django-redis
django-robots
django-timezone-field
django-utils-six
djongo
dnspython
et_xmlfile
firebase-admin
Flask
gcloud
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
graphene
graphene-django
graphql-core
graphql-relay
grpcio
grpcio-status
gspread
gunicorn
httplib2
idna
itsdangerous
Jinja2
jwcrypto
jws
kombu
MarkupSafe
msgpack
numpy
oauth2client
oauthlib
openpyxl
packaging
pandas
phonenumbers
pillow
promise
prompt_toolkit
proto-plus
protobuf
psycopg2
pyasn1
pyasn1_modules
pycparser
pycryptodome
PyJWT
pymongo
pyparsing
Pyrebase4
python-crontab
python-dateutil
python-dotenv
python-jwt
pytz
redis
requests
requests-oauthlib
requests-toolbelt
rsa
Rx
simplejson
singledispatch
six
sqlparse
tablib
text-unidecode
typing_extensions
tzdata
uritemplate
urllib3
vine
vue
wcwidth
Werkzeug
whitenoise
wrapt
xlrd